[FIX] renamed role bypass compute field
