+++ base_user_role_extended/views/res_role_views.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0"?>
+<?xml version="1.0" ?>
 <odoo>
     <record id="view_res_users_role_form_inherit" model="ir.ui.view">
         <field name="name">res.role.form.menu.inherit</field>
@@ -7,11 +7,7 @@
         <field name="arch" type="xml">
             <xpath expr="//notebook" position="inside">
                 <page string="Menu Items" id="menu_ids">
-                    <field
-                        name="menu_ids"
-                        context="{'ir.ui.menu.full_list': True}"
-                    >
-                    </field>
+                    <field name="menu_ids" context="{'ir.ui.menu.full_list': True}" />
                 </page>
             </xpath>
         </field>
